fix: removed few code smells
